[Fixes #9813] Duplicate keyword error raised on massive upload
