+++ geonode/base/models.py
@@ -27,6 +27,7 @@ import traceback
 from sequences.models import Sequence
 
 from sequences import get_next_value
+from django.db import transaction
 
 from django.db import models
 from django.db.models import Max
@@ -462,13 +463,20 @@ class _HierarchicalTagManager(_TaggableManager):
         tag_objs = set(tags) - str_tags
         # If str_tags has 0 elements Django actually optimizes that to not do a
         # query.  Malcolm is very smart.
-        existing = self.through.tag_model().objects.filter(
+        '''
+        To avoid concurrency with the keyword in case of a massive upload.
+        With the transaction block and the select_for_update,
+        we can easily handle the concurrency.
+        DOC: https://docs.djangoproject.com/en/3.2/ref/models/querysets/#select-for-update
+        '''
+        existing = self.through.tag_model().objects.select_for_update().filter(
             name__in=str_tags, **tag_kwargs
         )
-        tag_objs.update(existing)
-        new_ids = set()
-        for new_tag in str_tags - set(t.name for t in existing):
-            if new_tag:
+        with transaction.atomic():
+            tag_objs.update(existing)
+            new_ids = set()
+            _new_keyword = str_tags - set(t.name for t in existing)
+            for new_tag in list(_new_keyword):
                 new_tag = escape(new_tag)
                 new_tag_obj = HierarchicalKeyword.add_root(name=new_tag)
                 tag_objs.add(new_tag_obj)
