+++ geonode/tests/smoke.py
@@ -17,6 +17,9 @@
 #
 #########################################################################
 
+from unittest import TestCase
+from geonode.base.populate_test_data import create_single_dataset
+from geonode.resource.utils import metadata_storers
 from geonode.tests.base import GeoNodeBaseTestSupport
 
 import os
@@ -336,3 +339,61 @@ class UserMessagesTestCase(GeoNodeBaseTestSupport):
             response,
             f"{settings.SITEURL[:-1]}{account_login_url}?next=http%3A//testserver{target_url}"
         )
+
+
+'''
+Smoke test to explain how the new function for multiple storer will work
+Is required to define a fuction that takes 2 parametersand return 2 parameters.
+            Parameters:
+                    dataset: (Dataset): Dataset instance
+                    custom (dict): Custom dict generated by the parser
+
+            Returns:
+                    None
+'''
+
+
+class TestMetadataStorers(TestCase):
+    def setUp(self):
+        self.dataset = create_single_dataset('metadata-storer')
+        self.uuid = self.dataset.uuid
+        self.abstract = self.dataset.abstract
+        self.custom = {
+            "processes": {"uuid": "abc123cfde", "abstract": "updated abstract"},
+            "second-stage": {"title": "Updated Title", "abstract": "another update"},
+        }
+
+    @override_settings(METADATA_STORERS=['geonode.tests.smoke.dummy_metadata_storer'])
+    def test_will_use_single_storers_defined(self):
+        metadata_storers(self.dataset, self.custom)
+        self.assertEqual('abc123cfde', self.dataset.uuid)
+        self.assertEqual("updated abstract", self.dataset.abstract)
+
+    @override_settings(
+        METADATA_STORERS=[
+            "geonode.tests.smoke.dummy_metadata_storer",
+            "geonode.tests.smoke.dummy_metadata_storer2",
+        ]
+    )
+    def test_will_use_multiple_storers_defined(self):
+        dataset = metadata_storers(self.dataset, self.custom)
+        self.assertEqual('abc123cfde', dataset.uuid)
+        self.assertEqual("another update", dataset.abstract)
+        self.assertEqual("Updated Title", dataset.title)
+
+
+'''
+Just a dummy function required for the smoke test above
+'''
+
+
+def dummy_metadata_storer(dataset, custom):
+    if custom.get('processes', None):
+        for key, value in custom['processes'].items():
+            setattr(dataset, key, value)
+
+
+def dummy_metadata_storer2(dataset, custom):
+    if custom.get('second-stage', None):
+        for key, value in custom['second-stage'].items():
+            setattr(dataset, key, value)
