+++ geonode/upload/tests/test_utils.py
@@ -20,10 +20,7 @@
 """unit tests for geonode.upload.utils module"""
 
 from django.conf import settings
-from django.test.utils import override_settings
-from django.test import TestCase
 
-from geonode.base.populate_test_data import create_single_dataset
 from geonode.tests.base import GeoNodeBaseTestSupport
 from lxml import etree
 
@@ -82,130 +79,3 @@ class UtilsTestCase(GeoNodeBaseTestSupport):
             getattr(settings, "DEFAULT_MAX_UPLOAD_SIZE", 104857600)
         )
         upload_size.delete()
-
-
-class TestHandleMetadataKeyword(TestCase):
-    fixtures = [
-        "test_thesaurus.json"
-    ]
-
-    def setUp(self):
-        self.keyword = [
-            {
-                "keywords": ["features", "test_dataset"],
-                "thesaurus": {"date": None, "datetype": None, "title": None},
-                "type": "theme",
-            },
-            {
-                "keywords": ["no conditions to access and use"],
-                "thesaurus": {
-                    "date": "2020-10-30T16:58:34",
-                    "datetype": "publication",
-                    "title": "Test for ordering",
-                },
-                "type": None,
-            },
-            {
-                "keywords": ["ad", "af"],
-                "thesaurus": {
-                    "date": "2008-06-01",
-                    "datetype": "publication",
-                    "title": "GEMET - INSPIRE themes, version 1.0",
-                },
-                "type": None,
-            },
-            {"keywords": ["Global"], "thesaurus": {"date": None, "datetype": None, "title": None}, "type": "place"},
-        ]
-        self.dataset = create_single_dataset('keyword-handler')
-        self.sut = utils.KeywordHandler(
-            instance=self.dataset,
-            keywords=self.keyword
-        )
-
-    def test_return_empty_if_keywords_is_an_empty_list(self):
-        setattr(self.sut, 'keywords', [])
-        keyword, thesaurus_keyword = self.sut.handle_metadata_keywords()
-        self.assertListEqual([], keyword)
-        self.assertListEqual([], thesaurus_keyword)
-
-    def test_should_return_the_expected_keyword_extracted_from_raw_and_the_thesaurus_keyword(self):
-        keyword, thesaurus_keyword = self.sut.handle_metadata_keywords()
-        self.assertSetEqual({"features", "test_dataset", "no conditions to access and use"}, set(keyword))
-        self.assertListEqual(["ad", "af"], [x.alt_label for x in thesaurus_keyword])
-
-    def test_should_assign_correclty_the_keyword_to_the_dataset_object(self):
-        self.sut.set_keywords()
-        current_keywords = [keyword.name for keyword in self.dataset.keywords.all()]
-        current_tkeyword = [t.alt_label for t in self.dataset.tkeywords.all()]
-        self.assertSetEqual({"features", "test_dataset", "no conditions to access and use"}, set(current_keywords))
-        self.assertSetEqual({"ad", "af"}, set(current_tkeyword))
-
-    def test_is_thesaurus_available_should_return_queryset_with_existing_thesaurus(self):
-        keyword = "ad"
-        thesaurus = {"title": "GEMET - INSPIRE themes, version 1.0"}
-        actual = self.sut.is_thesaurus_available(thesaurus, keyword)
-        self.assertEqual(1, len(actual))
-
-    def test_is_thesaurus_available_should_return_empty_queryset_for_non_existing_thesaurus(self):
-        keyword = "ad"
-        thesaurus = {"title": "Random Thesaurus Title"}
-        actual = self.sut.is_thesaurus_available(thesaurus, keyword)
-        self.assertEqual(0, len(actual))
-
-
-'''
-Smoke test to explain how the new function for multiple storer will work
-Is required to define a fuction that takes 2 parametersand return 2 parameters.
-            Parameters:
-                    dataset: (Dataset): Dataset instance
-                    custom (dict): Custom dict generated by the parser
-
-            Returns:
-                    None
-'''
-
-
-class TestMetadataStorers(TestCase):
-    def setUp(self):
-        self.dataset = create_single_dataset('metadata-storer')
-        self.uuid = self.dataset.uuid
-        self.abstract = self.dataset.abstract
-        self.custom = {
-            "processes": {"uuid": "abc123cfde", "abstract": "updated abstract"},
-            "second-stage": {"title": "Updated Title", "abstract": "another update"},
-        }
-
-    @override_settings(METADATA_STORERS=['geonode.upload.tests.test_utils.dummy_metadata_storer'])
-    def test_will_use_single_storers_defined(self):
-        utils.metadata_storers(self.dataset, self.custom)
-        self.assertEqual('abc123cfde', self.dataset.uuid)
-        self.assertEqual("updated abstract", self.dataset.abstract)
-
-    @override_settings(
-        METADATA_STORERS=[
-            "geonode.upload.tests.test_utils.dummy_metadata_storer",
-            "geonode.upload.tests.test_utils.dummy_metadata_storer2",
-        ]
-    )
-    def test_will_use_multiple_storers_defined(self):
-        dataset = utils.metadata_storers(self.dataset, self.custom)
-        self.assertEqual('abc123cfde', dataset.uuid)
-        self.assertEqual("another update", dataset.abstract)
-        self.assertEqual("Updated Title", dataset.title)
-
-
-'''
-Just a dummy function required for the smoke test above
-'''
-
-
-def dummy_metadata_storer(dataset, custom):
-    if custom.get('processes', None):
-        for key, value in custom['processes'].items():
-            setattr(dataset, key, value)
-
-
-def dummy_metadata_storer2(dataset, custom):
-    if custom.get('second-stage', None):
-        for key, value in custom['second-stage'].items():
-            setattr(dataset, key, value)
